requirements.txt
asgiref==3.2.3
Click==7.0
Django==3.0.3
django-ckeditor==5.8.0
django-js-asset==1.2.2
Pillow==7.0.0
pyparsing==2.4.6
pytz==2019.3
sqlparse==0.3.0
svgwrite==1.3.1
Tree==0.2.4
